# api/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kubernetes RBAC Generator API",
        "description": "A REST API service for visualizing and analyzing Kubernetes Role-Based Access Control (RBAC) configurations. This service provides comprehensive insights into roles, bindings, and principals across your Kubernetes clusters.",
        "version": "1.0.0",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/djryanj/k8s-rbactory-backend",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Production server"
        }
    ],
...
    "paths": {
...
        "/counts": {
            "get": {
                "tags": [
                    "rbac"
                ],
                "summary": "Get resource counts",
                "description": "Returns counts of all RBAC resources in the cluster",
                "operationId": "getCounts",
                "responses": {
                    "200": {
                        "description": "Resource counts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceCounts"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/roles": {
            "get": {
                "tags": [
                    "rbac"
                ],
                "summary": "List roles",
                "description": "Returns a paginated list of Kubernetes roles, optionally filtered by namespace",
                "operationId": "listRoles",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/NamespaceQuery"
                    },
                    {
                        "$ref": "#/components/parameters/Limit"
                    },
                    {
                        "$ref": "#/components/parameters/Offset"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Roles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RBACList"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/roles/{namespace}/{name}": {
            "get": {
                "tags": [
                    "rbac"
                ],
                "summary": "Get role",
                "description": "Returns a specific Kubernetes role by namespace and name",
                "operationId": "getRole",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/NamespacePath"
                    },
                    {
                        "$ref": "#/components/parameters/NamePath"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RBACResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/clusterroles": {
            "get": {
                "tags": [
                    "rbac"
                ],
                "summary": "List cluster roles",
                "description": "Returns a paginated list of Kubernetes cluster roles",
                "operationId": "listClusterRoles",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Limit"
                    },
                    {
                        "$ref": "#/components/parameters/Offset"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cluster roles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RBACList"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/clusterroles/{name}": {
            "get": {
                "tags": [
                    "rbac"
                ],
                "summary": "Get cluster role",
                "description": "Returns a specific Kubernetes cluster role by name",
                "operationId": "getClusterRole",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/NamePath"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cluster role retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RBACResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/rolebindings": {
            "get": {
                "tags": [
                    "rbac"
                ],
                "summary": "List role bindings",
                "description": "Returns a paginated list of Kubernetes role bindings, optionally filtered by namespace",
                "operationId": "listRoleBindings",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/NamespaceQuery"
                    },
                    {
                        "$ref": "#/components/parameters/Limit"
                    },
                    {
                        "$ref": "#/components/parameters/Offset"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role bindings retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RBACList"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/clusterrolebindings": {
            "get": {
                "tags": [
                    "rbac"
                ],
                "summary": "List cluster role bindings",
                "description": "Returns a paginated list of Kubernetes cluster role bindings",
                "operationId": "listClusterRoleBindings",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Limit"
                    },
                    {
                        "$ref": "#/components/parameters/Offset"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cluster role bindings retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RBACList"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/principals": {
            "get": {
                "tags": [
                    "principals"
                ],
                "summary": "List principals",
                "description": "Returns a paginated list of all principals (users, groups, service accounts) with their associated bindings and roles",
                "operationId": "listPrincipals",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/NamespaceQuery"
                    },
                    {
                        "$ref": "#/components/parameters/Limit"
                    },
                    {
                        "$ref": "#/components/parameters/Offset"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Principals retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PrincipalList"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "responses": {
...
            "InternalServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        },
                        "example": {
                            "error": "Internal server error"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT token for authentication (if implemented)"
            }
        }
    },
    "security": []
}